Hotel-Management subproject updated from commit `a71b998` to `7ede8c4-dirty`.
This diff indicates that there has been a change in the commit history of the subproject `Hotel-Management`. The new commit `7ede8c4-dirty` is now being used in the project, replacing the old commit `a71b998`. The "dirty" suffix suggests that there may be uncommitted changes in the working directory of the subproject.